Fixed a warning in the vector add example.

+++ examples/vector_add.cu
@@ -83,7 +83,7 @@ int main() {
     C.copy_to(result);
 
     // Correctness check
-    for (size_t i = 0; i < n; i++) {
+    for (long i = 0; i < n; i++) {
         if (result[i] != float(i) + 1.0F) {
             std::cerr << "Wrong result at " << i << " : " << result[i] << " != " << float(i) + 1
                       << std::endl;
